Rename crypto_rand.rs to rand.rs
